Bootstrap menu bar app from app delegate
